# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "409": {
            "description": "Email already exists"
          }
        },
        "summary": "Register a new user",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "Login user",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/profile": {
      "get": {
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User profile retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/profile/name": {
      "put": {
        "operationId": "AuthController_updateName",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "John Doe",
                    "minLength": 5
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User name updated successfully"
          },
          "400": {
            "description": "Invalid name"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user name",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/profile/email": {
      "put": {
        "operationId": "AuthController_updateEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User email updated successfully"
          },
          "400": {
            "description": "Invalid email format"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Email already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user email",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/password/reset": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "user@example.com"
                  },
                  "newPassword": {
                    "type": "string",
                    "example": "newPassword123",
                    "minLength": 8
                  }
                },
                "required": [
                  "email",
                  "newPassword"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successfully"
          },
          "401": {
            "description": "User not found"
          }
        },
        "summary": "Reset user password",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/password/forgot": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "user@example.com"
                  }
                },
                "required": [
                  "email"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Temporary password sent to email if account exists"
          },
          "400": {
            "description": "Invalid email format"
          }
        },
        "summary": "Request password reset with temporary password",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/profile/notification-token": {
      "put": {
        "operationId": "AuthController_updateNotificationToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "notificationToken": {
                    "type": "string",
                    "example": "fcm_token_example_123456789abcdef",
                    "description": "Firebase Cloud Messaging token for push notifications"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notification token updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user notification token",
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/consumables": {
      "post": {
        "operationId": "ConsumablesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConsumableDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Consumable created successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new consumable",
        "tags": [
          "Consumables"
        ]
      },
      "get": {
        "operationId": "ConsumablesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Consumables retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all consumables",
        "tags": [
          "Consumables"
        ]
      }
    },
    "/consumables/low-stock": {
      "get": {
        "operationId": "ConsumablesController_findLowStock",
        "parameters": [
          {
            "name": "threshold",
            "required": false,
            "in": "query",
            "description": "Stock threshold (default: 5)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Low stock consumables retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get low stock consumables",
        "tags": [
          "Consumables"
        ]
      }
    },
    "/consumables/{id}": {
      "get": {
        "operationId": "ConsumablesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Consumable ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Consumable retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Consumable not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get consumable by ID",
        "tags": [
          "Consumables"
        ]
      },
      "patch": {
        "operationId": "ConsumablesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Consumable ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateConsumableDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Consumable updated successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Consumable not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update consumable",
        "tags": [
          "Consumables"
        ]
      },
      "delete": {
        "operationId": "ConsumablesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Consumable ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Consumable deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Consumable not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete consumable",
        "tags": [
          "Consumables"
        ]
      }
    },
...
  },
  "info": {
    "title": "ShelfSpot API",
    "description": "API for ShelfSpot application with JWT Authentication",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "password123"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "notificationToken": {
            "type": "string",
            "example": "fcm_token_example_123456"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "token_type": {
            "type": "string",
            "example": "bearer"
          },
          "expires_in": {
            "type": "number",
            "example": 3600
          },
          "user": {
            "type": "object",
            "example": {
              "id": "1",
              "email": "user@example.com",
              "name": "John Doe",
              "admin": false
            }
          }
        },
        "required": [
          "access_token",
          "token_type",
          "expires_in",
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "password123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "UpdateEmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "newemail@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
...
      "CreateConsumableDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Coffee Beans",
            "description": "Consumable name"
          },
          "quantity": {
            "type": "number",
            "example": 5,
            "description": "Consumable quantity",
            "minimum": 0
          },
          "status": {
            "type": "string",
            "example": "Available",
            "description": "Consumable status"
          },
          "roomId": {
            "type": "number",
            "example": 1,
            "description": "Room ID"
          },
          "placeId": {
            "type": "number",
            "example": 1,
            "description": "Place ID"
          },
          "containerId": {
            "type": "number",
            "example": 1,
            "description": "Container ID"
          }
        },
        "required": [
          "name",
          "quantity",
          "roomId",
          "placeId"
        ]
      },
      "UpdateConsumableDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Coffee Beans",
            "description": "Consumable name"
          },
          "quantity": {
            "type": "number",
            "example": 5,
            "description": "Consumable quantity",
            "minimum": 0
          },
          "status": {
            "type": "string",
            "example": "Available",
            "description": "Consumable status"
          },
          "roomId": {
            "type": "number",
            "example": 1,
            "description": "Room ID"
          },
          "placeId": {
            "type": "number",
            "example": 1,
            "description": "Place ID"
          },
          "containerId": {
            "type": "number",
            "example": 1,
            "description": "Container ID"
          }
        }
      },
...
    }
  }
}